Added the four basic types of connections, and the basic registration. Still need more code coverage and extend the unit tests.

+++ phenomena/connection_types/listener.py
@@ -5,7 +5,7 @@ from ontic.schema_type import SchemaType
 class Listener(OnticType):
     ONTIC_SCHEMA = SchemaType({
         'id': {
-            'type': int,
+            'type': long,
             'required': True,
         },
         'type': {
